refactor(docs): table dx
