+++ apps/docs/content/components/table/empty.raw.jsx
@@ -1,0 +1,14 @@
+import {Table, TableHeader, TableColumn, TableBody} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Table aria-label="Example empty table">
+      <TableHeader>
+        <TableColumn>NAME</TableColumn>
+        <TableColumn>ROLE</TableColumn>
+        <TableColumn>STATUS</TableColumn>
+      </TableHeader>
+      <TableBody emptyContent={"No rows to display."}>{[]}</TableBody>
+    </Table>
+  );
+}
